moved content component to main component for generic page layout
